GoogleSpeechAnalysis/requirements.txt
google-cloud-speech
pyaudio
numpy